requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
albumentations=1.3.1=pyhd8ed1ab_0
anyio=4.0.0=pypi_0
astor=0.8.1=pypi_0
attrdict=2.0.1=pypi_0
babel=2.13.1=pypi_0
bce-python-sdk=0.8.95=pypi_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
blinker=1.7.0=pypi_0
blosc=1.21.3=h6c2663c_0
bottleneck=1.3.5=py39h080aedc_0
brotli=1.0.9=ha925a31_2
brotli-python=1.0.9=py39hd77b12b_7
bzip2=1.0.8=h8ffe710_4
ca-certificates=2023.7.22=h56e8100_0
cachetools=5.3.2=pypi_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py39h2bbff1b_3
cfitsio=3.470=h2bbff1b_7
chardet=4.0.0=py39haa95532_1003
charls=2.2.0=h6c2663c_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=win_pyh7428d3b_0
cloudpickle=3.0.0=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cryptography=41.0.3=py39h89fc84f_0
cssselect=1.2.0=pypi_0
cssutils=2.9.0=pypi_0
cuda-cccl=12.3.52=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.12.1=pyhd8ed1ab_0
cython=3.0.5=pypi_0
cytoolz=0.12.0=py39h2bbff1b_0
dask-core=2023.11.0=pyhd8ed1ab_0
decorator=5.1.1=pypi_0
eigen=3.4.0=h2d74725_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.1.3=pypi_0
ffmpeg=4.2.3=ha925a31_0
filelock=3.9.0=py39haa95532_0
fire=0.5.0=pypi_0
flask=3.0.0=pypi_0
flask-babel=4.0.0=pypi_0
fonttools=4.44.0=pypi_0
freetype=2.10.4=h546665d_1
fsspec=2023.10.0=pyhca7485f_0
future=0.18.3=pypi_0
giflib=5.2.1=h8d14728_2
glib=2.69.1=h5dc1a3c_2
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
gtts=2.4.0=pyhd8ed1ab_0
h11=0.14.0=pypi_0
hdf5=1.10.6=nompi_he0bbb20_101
httpcore=1.0.2=pypi_0
httpx=0.25.1=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py39haa95532_0
imagecodecs=2021.8.26=py39h319e4f4_2
imageio=2.31.5=pyh8c1a49c_0
imgaug=0.4.0=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
importlib_resources=6.1.1=pyhd8ed1ab_0
intel-openmp=2023.2.0=h57928b3_50497
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=py39haa95532_0
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h8ffe710_2
jxrlib=1.1=h8ffe710_2
kiwisolver=1.4.5=pypi_0
lazy_loader=0.3=pyhd8ed1ab_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libclang=12.0.0=default_h627e005_2
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.4.52=0
libcurand-dev=10.3.4.52=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.2=h8ffe710_5
libiconv=1.17=h8ffe710_0
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libogg=1.3.4=h8ffe710_1
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=h0e60522_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.10.4=h0ad7f3c_1
libxslt=1.1.37=h2bbff1b_1
libzopfli=1.0.3=h0e60522_0
lmdb=1.4.1=pypi_0
locket=1.0.0=pyhd8ed1ab_0
lxml=4.9.3=pypi_0
lz4-c=1.9.4=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.3=pypi_0
matplotlib=3.8.1=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.8=py39h2bbff1b_0
mkl_random=1.2.4=py39h59b6b97_0
mpmath=1.3.0=py39haa95532_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
networkx=2.8.8=pyhd8ed1ab_0
numexpr=2.8.7=py39h2cd9be0_0
numpy=1.26.0=py39h055cbcc_0
numpy-base=1.26.0=py39h65a83cf_0
opencv=4.6.0=py39hf11a4ad_3
opencv-contrib-python=4.6.0.66=pypi_0
opencv-python=4.6.0.66=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openpyxl=3.1.2=pypi_0
openssl=3.0.12=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
paddleocr=2.7.0.3=pypi_0
paddlepaddle-gpu=2.5.2=pypi_0
pandas=2.1.1=py39h4ed8f06_0
partd=1.4.1=pyhd8ed1ab_0
pcre=8.45=h0e60522_0
pdf2docx=0.5.6=pypi_0
pillow=10.0.1=py39h045eedc_0
pip=23.3=py39haa95532_0
playsound=1.3.0=pypi_0
premailer=3.10.0=pypi_0
protobuf=3.20.2=pypi_0
psutil=5.9.6=pypi_0
py-cpuinfo=9.0.0=pyhd8ed1ab_0
pyclipper=1.3.0.post5=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.19.0=pypi_0
pymupdf=1.20.2=pypi_0
pyopenssl=23.2.0=py39haa95532_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.18=h1aa4202_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-docx=1.1.0=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=2.0.1=py3.9_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h16d0643_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py39haa95532_0
pywavelets=1.4.1=py39h2bbff1b_0
pyyaml=6.0=py39hb82d6ee_4
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=h5bd16bc_7
qtwebkit=5.212=h2bbfb41_5
qudida=0.0.4=pyhd8ed1ab_0
rapidfuzz=3.5.2=pypi_0
rarfile=4.1=pypi_0
requests=2.31.0=pypi_0
scikit-image=0.19.3=py39hd77b12b_1
scikit-learn=1.3.0=py39h4ed8f06_0
scipy=1.11.3=py39h309d312_0
seaborn=0.12.2=py39haa95532_0
setuptools=68.0.0=py39haa95532_0
shapely=2.0.2=pypi_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h6c2663c_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h2bbff1b_0
sympy=1.11.1=py39haa95532_0
tbb=2021.8.0=h59b6b97_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.2.0=pyha21a80b_0
tifffile=2021.4.8=pyhd8ed1ab_0
tk=8.6.12=h2bbff1b_0
toolz=0.12.0=pyhd8ed1ab_0
torchaudio=2.0.2=pypi_0
torchvision=0.15.2=pypi_0
tqdm=4.66.1=pyhd8ed1ab_0
typing-extensions=4.8.0=hd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h04d1e81_0
ultralytics=8.0.208=pyhf3ec037_0
urllib3=2.0.7=pypi_0
vc=14.2=h21ff451_1
visualdl=2.5.3=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=3.0.1=pyhd8ed1ab_0
wheel=0.41.2=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
xz=5.4.2=h8cc25b3_0
yaml=0.2.5=h8ffe710_2
zfp=0.5.5=h0e60522_8
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
